+++ modules/core/keeper/export_test.go
@@ -1,0 +1,9 @@
+package keeper
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+// ConvertToErrorEvents is a wrapper around convertToErrorEvents
+// to allow the function to be directly called in tests.
+func ConvertToErrorEvents(events sdk.Events) sdk.Events {
+	return convertToErrorEvents(events)
+}
